Add license
The Fedora SPEC file for KiCad, from which the SPEC files in this
repository have been derived, is available under the terms of the MIT
license. Include the text of the "Modern Style with sublicense" variant
of the MIT license (as outlined in the Fedora Contributor Agreement) in
this repository to be compliant to its terms.

All of the packaging scripts in this repository are made available under
the same variant of the MIT license.

https://fedoraproject.org/wiki/Licensing:Main#License_of_Fedora_SPEC_Files
https://fedoraproject.org/wiki/Legal:Fedora_Project_Contributor_Agreement
https://fedoraproject.org/wiki/Licensing:MIT#Modern_Style_with_sublicense