Fix application launchers
The "Development" category was recently changed to "Science". Make sure
to get rid of it to not duplicate the KiCad menu entries outside of the
"Electronics" category.

Register both the "normal" and the "nightly" MIME types in application
launchers, to make it easier for users to select the correct application
from the file manager's context menu when opening files.