Remove documentation symlink
KiCad now relies on the KICAD_DOCS build parameter to find the
documentation and help files, so that the symlink that was previously
required to point from the expected to the actual location is obsolete.
Also, drop the %{kicad_bindir} and %{kicad_docdir} macros in an attempt
to further simplify the SPEC file.